Move CH9329Comm into its own module (towards #13)
CH9329Comm and its _signed_byte helper move from communication.py into a new kvm_serial/utils/ch9329.py. communication.py keeps the DataComm ABC and list_serial_ports. CH9350Comm will land alongside in ch9350.py.

baseop.py and tests/utils/test_communication.py updated to import from the new module path. No behaviour change.
